Separate code into client/server folders
