+++ tictactoe/src/client/Celebration.tsx
