+++ tictactoe/src/client/Square.tsx
